+++ internal/evaluate/nnue/utils.go
@@ -1,12 +1,12 @@
 package nnue
 
-func screlu(x int) int {
-	y := clamp(int(x), 0, QA)
+func screlu(x int16) int32 {
+	y := int32(clamp(x, 0, QA))
 	return y * y
 }
 
 // helper function to clamp bonus between MAX_HISTORY
-func clamp(n, low, high int) int {
+func clamp(n, low, high int16) int16 {
 	if n <= low {
 		return low
 	}
